Moves Tentacat.Client to a separate file
